- add new mathmodel SpiralModel - add one more layer to IGUILayout - add moveView() , zoomView() - reworked method for colecting debug information (from base to end state) - changed render targen on EditorState::render() - Editor already layred with usage switch (current_View_Generator) as sepatator - in Editor we can manipulate with SpiralModel

+++ source/core/math/models noice/Noise_data.h
@@ -8,7 +8,7 @@ struct NoiceData {
 	int gridSize = 0;
 	float chanceToSpawnTree = 0.0f;
 	// data 4-gen 
-	int seed = 0;
+	unsigned long long seed = 0;
 	int offsetSeed = 0;
 	unsigned octaves = 0;
 	float amplifire = 0.f;
